requirements.txt
langchain
streamlit
python-dotenv
openai
os